Move misplaced files to the correct location - Move `src/lib/helpers/crypter.ts` to `src/app/api/__helpers__/` - Move `src/lib/helpers/response.ts` to `src/app/api/__helpers__/` - Update imports for the moved files to reflect the new location
